+++ dota_rating/rating/models.py
@@ -11,6 +11,6 @@ class Match(models.Model):
     deaths = models.IntegerField()
     assists = models.IntegerField()
     win = models.BooleanField()
-    rating_change = models.FloatField()
+    rating_change = models.IntegerField()
     match_time = models.DateTimeField()
     duration = models.IntegerField(help_text="Duration in seconds")
